!fixup Split path functions into Paths
