+++ src/papyrus-lang-vscode/src/debugger/PapyrusDebugAdapterDescriptorFactory.ts
@@ -21,7 +21,8 @@ import { ICreationKitInfoProvider } from '../CreationKitInfoProvider';
 import { IExtensionConfigProvider } from '../ExtensionConfigProvider';
 import { take } from 'rxjs/operators';
 import { IPapyrusDebugSession } from './PapyrusDebugSession';
-import { toCommandLineArgs, getGameIsRunning, getDebugToolPath } from '../Utilities';
+import { toCommandLineArgs, getGameIsRunning } from '../Utilities';
+import { getDebugToolPath } from '../Paths';
 import { IExtensionContext } from '../common/vscode/IocDecorators';
 import { IDebugSupportInstallService, DebugSupportInstallState } from './DebugSupportInstallService';
 import { ILanguageClientManager } from '../server/LanguageClientManager';
